Reorganize flexspi API into separate module
